fixing compatibility issues on this branch
