+++ pom.xml
@@ -266,6 +266,30 @@
                         <generateBackupPoms>false</generateBackupPoms>
                     </configuration>
                 </plugin>
+                <plugin>
+                    <groupId>com.github.siom79.japicmp</groupId>
+                    <artifactId>japicmp-maven-plugin</artifactId>
+                    <version>0.11.1</version>
+                    <configuration>
+                        <oldVersion>
+                            <dependency>
+                                <groupId>${project.groupId}</groupId>
+                                <artifactId>${project.artifactId}</artifactId>
+                                <version>${okta.sdk.pereviousVersion}</version>
+                                <type>jar</type>
+                            </dependency>
+                        </oldVersion>
+                        <parameter>
+                            <onlyModified>true</onlyModified>
+                            <breakBuildBasedOnSemanticVersioning>true</breakBuildBasedOnSemanticVersioning>
+                            <excludes>
+                                <exclue>com.okta.sdk.impl.ds.InternalDataStore#create(java.lang.String,com.okta.sdk.resource.Resource)</exclue>
+                                <exclue>com.okta.sdk.impl.ds.InternalDataStore#create(java.lang.String,com.okta.sdk.resource.Resource,java.lang.Class)</exclue>
+                                <exclue>com.okta.sdk.impl.ds.InternalDataStore#save(java.lang.String,com.okta.sdk.resource.Resource)</exclue>
+                            </excludes>
+                        </parameter>
+                    </configuration>
+                </plugin>
             </plugins>
         </pluginManagement>
         <plugins>
@@ -301,21 +325,6 @@
                     <plugin>
                         <groupId>com.github.siom79.japicmp</groupId>
                         <artifactId>japicmp-maven-plugin</artifactId>
-                        <version>0.10.0</version>
-                        <configuration>
-                            <oldVersion>
-                                <dependency>
-                                    <groupId>${project.groupId}</groupId>
-                                    <artifactId>${project.artifactId}</artifactId>
-                                    <version>${okta.sdk.pereviousVersion}</version>
-                                    <type>jar</type>
-                                </dependency>
-                            </oldVersion>
-                            <parameter>
-                                <onlyModified>true</onlyModified>
-                                <breakBuildBasedOnSemanticVersioning>true</breakBuildBasedOnSemanticVersioning>
-                            </parameter>
-                        </configuration>
                         <executions>
                             <execution>
                                 <id>japicmp</id>
